// src/app/model/Clientpartenaire.ts
export class Clientpartenaire{
  dateCreation:Date;
  idClient: string;
  idClientPartenaire: string;
  idGroupeClientPartenaire: string;
  idJournee: string;
  idPartenaire: string;
  isActif: number;
  isconnected: number;
  nbrAlimentationFidelite: number;
  nom: string;
  password: string;
  prenom: string;
  qrCodePartn: string;
  soldePartn: number;
  srcCreation: string;
  tokenNotification: string
}